workflows/flyte-v1/requirements_dbt.txt
dbt-oracle==1.7.7
flytekit
flytekitplugins-dbt
#protobuf==6.32.0